requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
Pillow>=8.4.0
pykitti>=1.0.1
tqdm>=4.64.0
numpy>=1.21.0
tensorboard>=2.12.0
